requirements.txt
geopy==2.3.0
numpy==1.21.6
pandas==1.3.5
cryptography==36.0.2
google-cloud==0.34.0
google-cloud-bigquery==3.4.1
google-auth==2.15.0
